aula2

+++ demo/src/main/java/APISpring/repositories/UsuarioRepository.java
@@ -1,11 +1,13 @@
 package APISpring.repositories;
 
+
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import APISpring.entities.Usuario;
 
-public interface UsuarioRepository  extends JpaRepository <Usuario, Long> {
+public interface UsuarioRepository  extends JpaRepository <Usuario, Integer> {
 	Optional<Usuario> findByEmail(String email);
 }
+
